+++ src/rules/status.rs
@@ -1,5 +1,6 @@
 //! Generic implementations for all purpose statuses.
 
+use crate::status::StatusDuration;
 use crate::util::Id;
 #[cfg(feature = "serialization")]
 use serde::{Deserialize, Serialize};
@@ -12,16 +13,16 @@ use std::hash::Hash;
 pub struct SimpleStatus<I, V> {
     id: I,
     effect: V,
-    duration: Option<u16>,
+    max_duration: Option<StatusDuration>,
 }
 
 impl<I, V: Copy> SimpleStatus<I, V> {
     /// Creates a new `SimpleStatus`.
-    pub fn new(id: I, effect: V, duration: Option<u16>) -> SimpleStatus<I, V> {
+    pub fn new(id: I, effect: V, max_duration: Option<StatusDuration>) -> SimpleStatus<I, V> {
         SimpleStatus {
             id,
             effect,
-            duration,
+            max_duration,
         }
     }
 
@@ -35,28 +36,10 @@ impl<I, V: Copy> SimpleStatus<I, V> {
         self.effect = effect;
     }
 
-    /// Returns the remaining duration of this status.
+    /// Returns the maximum duration of this status.
     /// `None` means infinite duration.
-    pub fn duration(&self) -> Option<u16> {
-        self.duration
-    }
-
-    /// Decreases the duration by one.
-    pub fn advance(&mut self) {
-        if let Some(duration) = self.duration {
-            if duration > 0 {
-                self.duration = Some(duration - 1);
-            }
-        }
-    }
-
-    /// Returns `true` if this status has ended.
-    pub fn finished(&self) -> bool {
-        if let Some(duration) = self.duration {
-            duration == 0
-        } else {
-            false
-        }
+    pub fn max_duration(&self) -> Option<StatusDuration> {
+        self.max_duration
     }
 }